Removed helper class from UIL and moved it to Helpers project
